backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# NeuroMemory (本地安装)
# pip install -e ../NeuroMemory
# 或者等 NeuroMemory 发布到 PyPI 后：
# neuromemory>=2.0.0

# LLM
openai==1.10.0  # DeepSeek 使用 OpenAI 兼容接口
httpx==0.26.0

# Embedding (本地模型)
sentence-transformers==2.3.1
torch>=2.0.0  # sentence-transformers 依赖

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.6.0
